Fix misunderstanding
