+++ iidm/iidm-impl/src/test/java/com/powsybl/iidm/network/impl/tck/BusTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019-2025, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -7,6 +7,23 @@
  */
 package com.powsybl.iidm.network.impl.tck;
 
+import com.powsybl.commons.PowsyblException;
+import com.powsybl.iidm.network.Bus;
+import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.tck.AbstractBusBreakerTest;
+import org.junit.jupiter.api.Test;
 
-class BusTest extends AbstractBusBreakerTest { }
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class BusTest extends AbstractBusBreakerTest {
+
+    @Test
+    void testFictitiousP0AndFictitiousQ0ForInvalidatedBusInBusBreaker() {
+        Network network = createTestNetwork();
+        Bus bus = network.getVoltageLevel("VL1").getBusView().getBus("VL1_0");
+        network.getSwitch("BR1").setOpen(true);
+        assertThrows(PowsyblException.class, bus::getFictitiousP0, "Bus has been invalidated");
+        assertThrows(PowsyblException.class, bus::getFictitiousQ0, "Bus has been invalidated");
+    }
+
+}
